move init tests to separate file
